refactor(host-application): isolate approval workflow boundaries
