+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/mod.rs
@@ -2,6 +2,7 @@ mod approval_context_service;
 mod approval_service;
 mod cleanup_plans;
 mod approval_support;
+mod builder_branch_service;
 mod builder_cleanup_service;
 mod direct_merge_workflow_service;
 mod document_service;
